+++ CHANGELOG.rst
@@ -4,11 +4,19 @@ Changelog
 *latest*
 --------
 
+- New routines:
+
+  - ``utils.get_h_field``: Small routine to calculate the magnetic field from
+    the electric field using Faraday's law.
+  - ``utils.get_receiver``: Small wrapper to interpolate a field at receiver
+    positions.
+
 - Re-implemented the possibility to define isotropic magnetic permeabilities in
   ``utils.Model``. Magnetic permeability is not tri-axially included in the
   solver currently; however, it would not be too difficult to include if there
   is a need.
 - CPU-graph added on top of RAM-graph.
+- Expand ``utils.Field`` to work with pickle/shelve.
 - Small bugfix introduced in ebd2c9d5: ``sc_cycle`` and ``lr_cycle`` was not
   updated any longer at the end of a cycle (only affected ``sslsolver=True``.